Moved the Logging.hpp to mars/interfaces
