+++ interfaces/manifest.xml
@@ -6,6 +6,5 @@
     <depend package="simulation/mars/scripts/cmake" />
     <depend package="simulation/mars/common/utils" />
     <depend package="simulation/mars/common/data_broker" />
-    <depend package="simulation/mars/common/logging" />
     <tags>needs_opt</tags>
 </package>
