+++ scene_loader/src/Load.cpp
@@ -38,7 +38,7 @@
 #include <mars/interfaces/sim/EntityManagerInterface.h>
 #include <mars/interfaces/sim/LoadSceneInterface.h>
 #include <mars/utils/misc.h>
-#include <mars/logging/Logging.hpp>
+#include <mars/interfaces/Logging.hpp>
 
 //#define DEBUG_PARSE_SENSOR 1
