+++ sim/src/core/NodeManager.cpp
@@ -35,7 +35,7 @@
 #include <mars/interfaces/sim/SimulatorInterface.h>
 #include <mars/interfaces/graphics/GraphicsManagerInterface.h>
 #include <mars/interfaces/terrainStruct.h>
-#include <mars/logging/Logging.hpp>
+#include <mars/interfaces/Logging.hpp>
 
 #include <mars/interfaces/utils.h>
 #include <mars/utils/mathUtils.h>
