+++ sim/src/core/SensorManager.cpp
@@ -57,7 +57,7 @@
 
 #include <mars/interfaces/sim/SimulatorInterface.h>
 #include <mars/utils/MutexLocker.h>
-#include <mars/logging/Logging.hpp>
+#include <mars/interfaces/Logging.hpp>
 
 #include <cstdio>
 #include <stdexcept>
